+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Based on Realm JS v10.22.0: See changelog below for details on enhancements and 
    * `Realm.Credentials#serverApiKey`
    * `Realm.Credentials#userApiKey`
 * When no object is found calling `Realm#objectForPrimaryKey`, `null` is returned instead of `undefined`
+* Replaced `Realm#empty` with `Realm#isEmpty`
+* Replaced `Realm#readOnly` with `Realm#isReadOnly`
 
 ### Enhancements
 * None